requirements.txt
bcrypt==3.1.7
cffi==1.14.1
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
Flask-Mail==0.9.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pip==20.2.2
pycparser==2.20
setuptools==49.6.0
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
WTForms==2.3.3
wheel
PyMySQL
pyjwt
